+++ file_test.go
@@ -477,39 +477,6 @@ func TestFilePosixSymlink(t *testing.T) {
 	}
 }
 
-func TestFilePosixWriteFile(t *testing.T) {
-	perm := os.FileMode(0o666)
-	want := fixPosixPerms(perm, false)
-
-	name, err := tempName(t)
-	if err != nil {
-		t.Fatal(err)
-
-		return
-	}
-
-	err = compat.WriteFile(name, helloBytes, want)
-	if err != nil {
-		t.Fatal(err)
-
-		return
-	}
-
-	fi, err := os.Stat(name)
-	if err != nil {
-		t.Fatal(err)
-
-		return
-	}
-
-	got := fi.Mode().Perm()
-	if got != want {
-		t.Fatalf("got 0%03o (%v), want 0%03o (%v)", got, got, want, want)
-
-		return
-	}
-}
-
 func TestFilePosixChmodInvalid(t *testing.T) {
 	err := compat.Chmod(invalidName, compat.CreatePerm)
 	if err == nil {
@@ -591,13 +558,6 @@ func TestFilePosixSymlinkInvalidNew(t *testing.T) {
 
 	new := invalidName
 	err = compat.Symlink(old, new)
-	if err == nil {
-		t.Fatal("got nil, want an error")
-	}
-}
-
-func TestFilePosixWriteFileInvalid(t *testing.T) {
-	err := compat.WriteFile(invalidName, helloBytes, compat.CreatePerm)
 	if err == nil {
 		t.Fatal("got nil, want an error")
 	}