+++ writefile_atomic_test.go
@@ -23,7 +23,9 @@ func TestWriteFileAtomic(t *testing.T) {
 
 	cleanup(t, file)
 
-	err = compat.WriteFileAtomic(file, helloBytes)
+	perm := compat.CreatePerm // 0o666
+	opts := []compat.Option{compat.WithAtomicity(true)}
+	err = compat.WriteFile(file, helloBytes, perm, opts...)
 	if err != nil {
 		fatalf(t, "Failed to write file: %q: %v", file, err)
 
@@ -35,7 +37,6 @@ func TestWriteFileAtomic(t *testing.T) {
 		t.Fatalf("Failed to stat file: %q: %v", file, err)
 	}
 
-	perm := compat.CreateTempPerm // 0o600
 	want := fixPerms(perm, false)
 	got := fi.Mode().Perm()
 	if got != want {
@@ -53,7 +54,9 @@ func TestWriteFileAtomicCurrentDir(t *testing.T) {
 
 	cleanup(t, file)
 
-	err = compat.WriteFileAtomic(base, helloBytes)
+	perm := compat.CreatePerm // 0o666
+	opts := []compat.Option{compat.WithAtomicity(true)}
+	err = compat.WriteFile(base, helloBytes, perm, opts...)
 	if err != nil {
 		fatalf(t, "Failed to write file: %q: %v", file, err)
 
@@ -65,7 +68,6 @@ func TestWriteFileAtomicCurrentDir(t *testing.T) {
 		t.Fatalf("Failed to stat file: %q: %v", file, err)
 	}
 
-	perm := compat.CreateTempPerm // 0o600
 	want := fixPerms(perm, false)
 	got := fi.Mode().Perm()
 	if got != want {
@@ -73,7 +75,7 @@ func TestWriteFileAtomicCurrentDir(t *testing.T) {
 	}
 }
 
-func TestWriteFileAtomicDefaultFileMode(t *testing.T) {
+func TestWriteFileAtomicNoPerms(t *testing.T) {
 	file, err := tempName(t)
 	if err != nil {
 		t.Fatalf("Failed to create temp file: %v", err)
@@ -81,7 +83,40 @@ func TestWriteFileAtomicDefaultFileMode(t *testing.T) {
 
 	cleanup(t, file)
 
-	err = compat.WriteFileAtomic(file, helloBytes, compat.WithDefaultFileMode(perm644))
+	perm := compat.CreatePerm // 0o600
+	opts := []compat.Option{compat.WithAtomicity(true)}
+	err = compat.WriteFile(file, helloBytes, 0, opts...)
+	if err != nil {
+		fatalf(t, "Failed to write file: %q: %v", file, err)
+
+		return // Tinygo doesn't support T.Fatal
+	}
+
+	fi, err := compat.Stat(file)
+	if err != nil {
+		t.Fatalf("Failed to stat file: %q: %v", file, err)
+	}
+
+	want := fixPerms(perm, false)
+	got := fi.Mode().Perm()
+	if got != want {
+		t.Fatalf("got %04o, want %04o", got, want)
+	}
+}
+
+func TestWriteFileAtomicWithDefaultFileMode(t *testing.T) {
+	file, err := tempName(t)
+	if err != nil {
+		t.Fatalf("Failed to create temp file: %v", err)
+	}
+
+	cleanup(t, file)
+
+	opts := []compat.Option{
+		compat.WithAtomicity(true),
+		compat.WithDefaultFileMode(perm644),
+	}
+	err = compat.WriteFile(file, helloBytes, 0, opts...)
 	if err != nil {
 		t.Fatalf("Failed to write file: %q: %v", file, err)
 	}
@@ -105,7 +140,7 @@ func TestWriteFileAtomicDefaultFileMode(t *testing.T) {
 		t.Fatalf("Failed to change file mode: %q: %v", file, err)
 	}
 
-	err = compat.WriteFileAtomic(file, helloBytes, compat.WithDefaultFileMode(perm644))
+	err = compat.WriteFile(file, helloBytes, 0, opts...)
 	if err != nil {
 		t.Fatalf("Failed to write file: %q: %v", file, err)
 	}
@@ -124,7 +159,7 @@ func TestWriteFileAtomicDefaultFileMode(t *testing.T) {
 	}
 }
 
-func TestWriteFileAtomicKeepFileMode(t *testing.T) { //nolint:dupl
+func TestWriteFileAtomicWithKeepFileMode(t *testing.T) { //nolint:dupl
 	file, err := tempName(t)
 	if err != nil {
 		t.Fatalf("Failed to create temp file: %v", err)
@@ -139,7 +174,11 @@ func TestWriteFileAtomicKeepFileMode(t *testing.T) { //nolint:dupl
 		t.Fatalf("Failed to create file: %q: %v", file, err)
 	}
 
-	err = compat.WriteFileAtomic(file, helloBytes, compat.WithKeepFileMode(true))
+	opts := []compat.Option{
+		compat.WithAtomicity(true),
+		compat.WithKeepFileMode(true),
+	}
+	err = compat.WriteFile(file, helloBytes, 0, opts...)
 	if err != nil {
 		t.Fatalf("Failed to write file: %q: %v", file, err)
 	}
@@ -156,7 +195,7 @@ func TestWriteFileAtomicKeepFileMode(t *testing.T) { //nolint:dupl
 	}
 }
 
-func TestWriteFileAtomicKeepFileModeFalse(t *testing.T) { //nolint:dupl
+func TestWriteFileAtomicWithKeepFileModeFalse(t *testing.T) { //nolint:dupl
 	file, err := tempName(t)
 	if err != nil {
 		t.Fatalf("Failed to create temp file: %v", err)
@@ -171,7 +210,11 @@ func TestWriteFileAtomicKeepFileModeFalse(t *testing.T) { //nolint:dupl
 		t.Fatalf("Failed to create file: %q: %v", file, err)
 	}
 
-	err = compat.WriteFileAtomic(file, helloBytes, compat.WithKeepFileMode(false))
+	opts := []compat.Option{
+		compat.WithAtomicity(true),
+		compat.WithKeepFileMode(false),
+	}
+	err = compat.WriteFile(file, helloBytes, 0, opts...)
 	if err != nil {
 		t.Fatalf("Failed to write file: %q: %v", file, err)
 	}
@@ -201,7 +244,11 @@ func TestWriteFileAtomicWithFileMode(t *testing.T) { //nolint:dupl
 
 	cleanup(t, file)
 
-	err = compat.WriteFileAtomic(file, helloBytes, compat.WithFileMode(perm644))
+	opts := []compat.Option{
+		compat.WithAtomicity(true),
+		compat.WithFileMode(perm644),
+	}
+	err = compat.WriteFile(file, helloBytes, 0, opts...)
 	if err != nil {
 		t.Fatalf("Failed to write file: %q: %v", file, err)
 	}
@@ -223,7 +270,7 @@ func TestWriteFileAtomicWithFileMode(t *testing.T) { //nolint:dupl
 		t.Fatalf("Failed to change file mode: %q: %v", file, err)
 	}
 
-	err = compat.WriteFileAtomic(file, helloBytes, compat.WithFileMode(perm644))
+	err = compat.WriteFile(file, helloBytes, 0, opts...)
 	if err != nil {
 		t.Fatalf("Failed to write file: %q: %v", file, err)
 	}
@@ -239,7 +286,7 @@ func TestWriteFileAtomicWithFileMode(t *testing.T) { //nolint:dupl
 	}
 }
 
-func TestWriteFileAtomicReadOnlyModeReset(t *testing.T) {
+func TestWriteFileAtomicWithReadOnlyModeReset(t *testing.T) {
 	if !compat.IsWindows {
 		skip(t, "Skipping test: requires Windows")
 
@@ -253,7 +300,12 @@ func TestWriteFileAtomicReadOnlyModeReset(t *testing.T) {
 
 	cleanup(t, file)
 
-	err = compat.WriteFileAtomic(file, helloBytes, compat.WithFileMode(perm400), compat.WithReadOnlyMode(compat.ReadOnlyModeReset))
+	opts := []compat.Option{
+		compat.WithAtomicity(true),
+		compat.WithFileMode(perm400),
+		compat.WithReadOnlyMode(compat.ReadOnlyModeReset),
+	}
+	err = compat.WriteFile(file, helloBytes, 0, opts...)
 	if err != nil {
 		t.Fatalf("Failed to write file: %q: %v", file, err)
 	}
@@ -270,14 +322,17 @@ func TestWriteFileAtomicReadOnlyModeReset(t *testing.T) {
 	}
 }
 
+// Invalid tests
+
 func TestWriteFileAtomicInvalid(t *testing.T) {
-	err := compat.WriteFileAtomic(invalidName, helloBytes)
+	opts := []compat.Option{compat.WithAtomicity(true)}
+	err := compat.WriteFile(invalidName, helloBytes, 0, opts...)
 	if err == nil {
 		t.Fatalf("got nil, want an error")
 	}
 }
 
-func TestWriteFileAtomicCantRead(t *testing.T) {
+func TestWriteFileAtomicInvalidCantRead(t *testing.T) {
 	file, err := tempFile(t)
 	if err != nil {
 		t.Fatalf("Failed to create temp file: %v", err)
@@ -297,15 +352,19 @@ func TestWriteFileAtomicCantRead(t *testing.T) {
 		t.Fatalf("Chmod: %v", err)
 	}
 
-	err = compat.WriteFileAtomic(file, helloBytes, compat.WithKeepFileMode(true))
+	opts := []compat.Option{
+		compat.WithAtomicity(true),
+		compat.WithKeepFileMode(true),
+	}
+	err = compat.WriteFile(file, helloBytes, 0, opts...)
 	if err != nil {
-		fatalf(t, "WriteFileAtomic: %v", err)
+		fatalf(t, "WriteFile: %v", err)
 
 		return // Tinygo doesn't support T.Fatal
 	}
 }
 
-func TestWriteFileAtomicReadOnlyDirectory(t *testing.T) {
+func TestWriteFileAtomicInvalidReadOnlyDirectory(t *testing.T) {
 	isRoot, _ := compat.IsRoot()
 	if isRoot && !compat.IsWindows {
 		skipf(t, "Skipping test: doesn't fail when root")
@@ -341,7 +400,8 @@ func TestWriteFileAtomicReadOnlyDirectory(t *testing.T) {
 		return
 	}
 
-	err = compat.WriteFileAtomic(file, helloBytes)
+	opts := []compat.Option{compat.WithAtomicity(true)}
+	err = compat.WriteFile(file, helloBytes, 0, opts...)
 	if err == nil {
 		fatal(t, "got nil, want an error")